Security fixes (#273)
- guard against malformed tarball headers
- guard against negative bytes read causing overflow during transfer of
  files by the kube go api
- fixed several logging messages
- updated lint and security tooling
- update kubernetes api libraries
- removed unused variables in functions

+++ cmd/root/collection/capture.go
@@ -116,7 +116,7 @@ func extractJobProgressText(line string) (status string, statusUX string, messag
 
 // Capture collects diagnostics, conf files and log files from the target hosts. Failures are permissive and
 // are first logged and then returned at the end with the reason for the failure.
-func StartCapture(c HostCaptureConfiguration, localDDCPath, localDDCYamlPath string, skipRESTCollect bool, disableFreeSpaceCheck bool, minFreeSpaceGB int) error {
+func StartCapture(c HostCaptureConfiguration, localDDCPath, localDDCYamlPath string, skipRESTCollect bool, disableFreeSpaceCheck bool, minFreeSpaceGB uint64) error {
 	host := c.Host
 	nodeState := consoleprint.NodeState{
 		Node:     host,
